refactor(baselines): hoist universal compat axes into envelope.js shared validator (resolves #2492)
